backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.6
langchain-openai==0.1.13
langchain-community==0.2.6
faiss-cpu==1.8.0
pypdf==4.2.0
python-dotenv==1.0.1
pydantic==2.7.4
streamlit==1.36.0
requests==2.32.3